# swagger.yaml
openapi: 3.0.0
info:
  title: Cirkle API
  version: 1.0.0
  description: |
    ## Cirkle — Social Events Platform API

    Cirkle is an AI-native social discovery platform for people to find their tribe and attend events together.

    ### Authentication
    All endpoints except `GET /cities`, `GET /lifestyle-tags`, and `POST /auth/login` require a Bearer token.

    Include the token in every protected request:
    Authorization: Bearer <your_jwt_token>

    ### Price Format
    All prices are in **paise** (₹1 = 100 paise). Example: ₹500 = `50000`.

    ### Phone Format
    Indian mobile numbers only. Format: `+91XXXXXXXXXX` where the 10-digit number starts with 6, 7, 8, or 9.
servers:
  - url: http://localhost:3000
    description: Development server
security:
  - bearerAuth: []
...
paths:
...
  /payments/orders:
    post:
      tags: [Payments]
      summary: Create order and start checkout
      description: |
        Claims a seat hold and creates a Razorpay order. The amount is computed server-side from the event price, an optional coupon, and the current GST rate — never accepted from the client.

        If the user already holds a live hold for this event, the existing order is **resumed** (same Razorpay order, same frozen price) rather than duplicated. `resumed: true` indicates this.

        The full price breakdown is frozen onto the order at creation and returned here.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [eventId, eventTicketCategoryId]
              properties:
                eventTicketCategoryId:
                  type: string
                  format: uuid
                  description: >-
                    Which ticket category the buyer chose — the `id` from
                    ticketCategories on GET /events/{id}. Required: the price
                    charged and the capacity consumed both come from it.
                eventId:
                  type: string
                  format: uuid
                couponCode:
                  type: string
                  example: CIRKLE100
                  description: Optional
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderCreated'
        '200':
          description: Existing live hold resumed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderCreated'
        '400':
          description: Missing event, or coupon became invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: |
            Blocked by a gate. Two distinct shapes — branch on the `error` field, not the prose:

            - `social_handles_required` — the event requires social handles the user hasn't supplied. Carries a `missing` array. This gate runs FIRST, before the invitation and capacity checks.
            - otherwise — invite-only event, an accepted invitation is required.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/SocialHandlesRequired'
                  - $ref: '#/components/schemas/Error'
              examples:
                socialHandlesRequired:
                  summary: Missing required social handles
                  value:
                    error: social_handles_required
                    missing: [instagram, linkedin]
                    message: This event requires your instagram, linkedin handles.
                invitationRequired:
                  summary: Invite-only event, no accepted invitation
                  value:
                    error: An accepted invitation is required to buy a ticket to this event
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: |
            Cannot buy. Branch on `error`:

            - `category_sold_out` — the chosen tier has no room for one more ticket of its size, or has zero stock. Unlimited tiers never return this.
            - `not_available_for_sale` — the event has no ticket categories configured, so nothing is purchasable.
            - otherwise — already ticketed, or the event has started.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                alreadyTicketed:
                  value:
                    error: You already have a ticket to this event
                soldOut:
                  value:
                    error: This event is sold out
                started:
                  value:
                    error: This event has already started
        '500':
          $ref: '#/components/responses/InternalError'
...
  /payments/orders/{id}:
    get:
      tags: [Payments]
      summary: Poll order status
      description: |
        Returns the current status of an order, scoped to the owner. The client polls this when the verify relay fails, waiting for the webhook to confirm the payment independently.

        `bookingRef` is null until the ticket exists. The signal to stop polling is `status: paid`.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Order status
          content:
            application/json:
              schema:
                type: object
                properties:
                  orderId:
                    type: string
                    format: uuid
                  status:
                    type: string
                    enum: [created, paid, failed, expired, refunded]
                  totalPaise:
                    type: integer
                  eventId:
                    type: string
                    format: uuid
                  bookingRef:
                    type: string
                    nullable: true
                    example: CRKL-7F3A2C9E
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /admin/admins:
    post:
      tags: [Admin — Admins]
      summary: Create an admin account
      description: |
        Creates a new admin (business_development or administrative). This is how administrative admins create BD team accounts.

        Requires the manage_admins capability — a business_development admin gets 403.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, displayName, role]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                displayName:
                  type: string
                role:
                  type: string
                  enum: [administrative, business_development]
      responses:
        '201':
          description: Admin created
          content:
            application/json:
              schema:
                type: object
                properties:
                  admin:
                    $ref: '#/components/schemas/Admin'
        '400':
          description: Validation error (email format, password length, missing displayName, invalid role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Caller lacks the manage_admins capability (business_development role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: You do not have permission to manage admin accounts
        '409':
          description: An admin with this email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      tags: [Admin — Admins]
      summary: List all admins
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: List of admin accounts
          content:
            application/json:
              schema:
                type: object
                properties:
                  admins:
                    type: array
                    items:
                      $ref: '#/components/schemas/Admin'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Caller lacks the manage_admins capability
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/admins/{id}:
    get:
      tags: [Admin — Admins]
      summary: Get a single admin account
      description: Same shape as the list rows. password_hash is never returned.
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: The admin
          content:
            application/json:
              schema:
                type: object
                properties:
                  admin:
                    $ref: '#/components/schemas/Admin'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Caller lacks the manage_admins capability
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Admin not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags: [Admin — Admins]
      summary: Update an admin account
      description: |
        Partial update — only fields present in the body are changed. Commonly used to deactivate an admin (`isActive: false`), which invalidates their existing token immediately (authenticateAdmin re-checks is_active on every request, not just at token issuance).

        **Lockout protection.** Two rules make admin management unbrickable. Both return 409 with a machine-readable `error` so the client can say why:

        - `cannot_deactivate_self` — nobody may deactivate their own account, unconditionally, however many other admins exist. The failure mode is losing the session you are in.
        - `last_administrative_admin` — the population of ACTIVE administrative admins may never reach zero, whether by deactivating the last one or demoting them to business_development.

        Self-demotion is deliberately NOT blanket-blocked: it is governed by the second rule alone, so an administrative admin may hand over and step down to BD as long as somebody else remains administrative.

        The population count is taken under a row lock in the same transaction as the update, so two admins demoting each other simultaneously cannot both pass and leave zero.

        **Password reset.** Send `password` to reset it — alone, or alongside other fields. A body containing only `password` is valid and no longer returns "No fields to update". The value is bcrypt-hashed, never returned and never logged; the audit row records only that a reset occurred, not the value. A reset changes no counts, so the lockout rules never block it — resetting your own password is always allowed.
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName:
                  type: string
                role:
                  type: string
                  enum: [administrative, business_development]
                isActive:
                  type: boolean
                password:
                  type: string
                  minLength: 8
                  format: password
                  description: >-
                    New password. Hashed with bcrypt before storage, never
                    returned in any response and never written to logs. May be
                    the only field in the body.
      responses:
        '200':
          description: Admin updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  admin:
                    $ref: '#/components/schemas/Admin'
        '400':
          description: Validation error, or no fields provided to update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Caller lacks the manage_admins capability
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Admin not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: |
            Blocked by lockout protection. Branch on `error`:

            - `cannot_deactivate_self` — you tried to deactivate your own account.
            - `last_administrative_admin` — the change would leave zero active administrative admins.
            - `concurrent_admin_change` — a competing admin change collided; retry.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum: [cannot_deactivate_self, last_administrative_admin, concurrent_admin_change]
                  message:
                    type: string
                    description: Human-readable. Branch on `error`, not on this.
              examples:
                self:
                  value:
                    error: cannot_deactivate_self
                    message: You cannot deactivate your own account.
                last:
                  value:
                    error: last_administrative_admin
                    message: There must be at least one active administrative admin.
        '500':
          $ref: '#/components/responses/InternalError'
...
  /admin/orders:
    get:
      tags: [Admin — Orders]
      summary: List orders
      description: |
        **PII masking by role.** End-user `phone` and `email` are role-gated. An `administrative` admin holds `view_pii` and receives the real values; a `business_development` admin receives MASKED ones — phone as `+91******210` (country code plus last three digits), email as `j****@gmail.com` (first character plus the domain, which is kept for context).

        Masking is applied server-side, so the real value never enters the response body for a BD admin and cannot be recovered from the network tab. There is no reveal endpoint or query flag.
        Filterable and combinable. Includes the full frozen price breakdown plus light event/user context for support lookup.

        **Date filtering.** `from` and `to` are both INCLUSIVE and are matched against `created_at`.

        A bare date (`YYYY-MM-DD`) means that whole day **in UTC** — `created_at` is timestamptz and the database session runs in UTC. So `from=2026-07-28&to=2026-07-28` returns everything created on the 28th UTC. (Internally `to` is applied as `created_at < to + 1 day`, which is exact at any timestamp precision.)

        A value carrying an explicit time component is honoured exactly as given and is NOT widened to end-of-day. Callers wanting an IST business day should send full UTC instants for its boundaries — which is what the admin portal already does, and it continues to work unchanged.
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [created, paid, failed, expired, refunded]
          required: false
        - in: query
          name: eventId
          schema:
            type: string
            format: uuid
          required: false
        - in: query
          name: userId
          schema:
            type: string
            format: uuid
          required: false
        - in: query
          name: from
          schema:
            type: string
            format: date
          required: false
          description: Lower bound on created_at (inclusive)
        - in: query
          name: to
          schema:
            type: string
            format: date
          required: false
          description: Upper bound on created_at (inclusive)
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Paginated list of orders
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginationEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/AdminOrder'
        '400':
          description: Invalid status value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/orders/{id}:
    get:
      tags: [Admin — Orders]
      summary: Get full order detail
      description: |
        Everything from the list view plus the coupon used (if any) and the linked ticket (if the order is paid).

        **PII masking by role.** End-user `phone` and `email` are role-gated: `administrative` admins hold `view_pii` and see real values, `business_development` admins receive masked ones (`+91******210`, `j****@gmail.com`). Masking is server-side, so the real value never enters the response body.
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Order detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  order:
                    allOf:
                      - $ref: '#/components/schemas/AdminOrder'
                      - type: object
                        properties:
                          coupon:
                            type: object
                            nullable: true
                            properties:
                              id:
                                type: string
                                format: uuid
                              code:
                                type: string
                              discountFlatPaise:
                                type: integer
                          ticket:
                            type: object
                            nullable: true
                            properties:
                              id:
                                type: string
                                format: uuid
                              bookingRef:
                                type: string
                              checkedIn:
                                type: boolean
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from POST /auth/login. Valid for 7 days.
  parameters:
    Limit:
      in: query
      name: limit
      schema:
        type: integer
        default: 50
        maximum: 100
        minimum: 1
      required: false
      description: Max rows to return. Default 50, clamped to 100 even if a larger value is requested.
    Offset:
      in: query
      name: offset
      schema:
        type: integer
        default: 0
        minimum: 0
      required: false
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: Error message here
    Admin:
      type: object
      description: password_hash is never included in any response.
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          example: admin@cirkle.live
        displayName:
          type: string
          example: Founder Admin
        role:
          type: string
          enum: [administrative, business_development]
          description: Only capability difference is manage_admins — administrative has it, business_development doesn't. Otherwise identical access.
        isActive:
          type: boolean
          description: Kill switch. A deactivated admin's existing token stops working immediately, not just at expiry. Absent from the login response (implied true — login already checked it).
        createdAt:
          type: string
          format: date-time
...
    PaginationEnvelope:
      type: object
      description: Shared shape for every list endpoint. `data` is added per-endpoint via allOf.
      properties:
        total:
          type: integer
          description: Total matching rows, ignoring limit/offset — for "showing X-Y of Z" and page controls
          example: 3847
        limit:
          type: integer
          example: 50
        offset:
          type: integer
          example: 100
    AdminOrder:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [created, paid, failed, expired, refunded]
        breakdown:
          $ref: '#/components/schemas/PriceBreakdown'
        paymentMethod:
          type: string
          nullable: true
          example: upi
        paymentMethodDetail:
          type: string
          nullable: true
          example: PhonePe
        razorpayOrderId:
          type: string
        razorpayPaymentId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        event:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
        user:
          type: object
          description: PII — phone is unmasked for now. See Admin — Users schema notes.
          properties:
            id:
              type: string
              format: uuid
            firstName:
              type: string
              nullable: true
            phone:
              type: string
...
    SocialHandlesRequired:
      type: object
      description: >-
        The 403 returned by POST /payments/orders and POST /events/{id}/invitations
        when the event requires social handles the user hasn't supplied.

        Branch on `error` — it distinguishes this from every other 403 on those
        endpoints (an invite-only event needing an accepted invitation returns a
        different body). Use `missing` to decide which fields the handle dialog
        should show. Never parse `message`.
      properties:
        error:
          type: string
          enum: [social_handles_required]
        missing:
          type: array
          description: Exactly the required handles the user lacks, in a stable order.
          items:
            type: string
            enum: [facebook, instagram, linkedin]
          example: [instagram, linkedin]
        message:
          type: string
          description: Human-readable, for logging or a fallback. Not a stable contract.
...
    PriceBreakdown:
      type: object
      description: All amounts in paise. GST is computed on the post-discount subtotal.
      properties:
        basePricePaise:
          type: integer
          example: 50000
        discountPaise:
          type: integer
          example: 10000
          description: Clamped to base price — never exceeds it
        subtotalPaise:
          type: integer
          example: 40000
          description: base − discount
        gstPercentage:
          type: number
          example: 18
        gstPaise:
          type: integer
          example: 7200
          description: Rounded to the nearest paisa
        totalPaise:
          type: integer
          example: 47200
          description: What Razorpay charges
    OrderCreated:
      type: object
      properties:
        orderId:
          type: string
          format: uuid
          description: Cirkle's order ID — used for status polling
        razorpayOrderId:
          type: string
          example: order_Nx8kL2mP9qR3sT
          description: Razorpay's order ID — passed to the checkout SDK
        razorpayKeyId:
          type: string
          example: rzp_test_XXXXXXXX
          description: Public key ID for the checkout SDK
        amount:
          type: integer
          example: 47200
          description: Total in paise, equals breakdown.totalPaise
        currency:
          type: string
          example: INR
        expiresAt:
          type: string
          format: date-time
          description: When the seat hold lapses
        resumed:
          type: boolean
          description: True if an existing live hold was returned instead of a new order
        breakdown:
          $ref: '#/components/schemas/PriceBreakdown'
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: No token provided
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Internal server error
...